requirements.txt
arrow==1.3.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
redis==5.1.0
rq==1.16.2
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20240906
tzdata==2024.1
